backend/requirements.txt
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.30.0,<1.0.0
httpx>=0.28.1,<1.0.0
pydantic>=2.9.0,<3.0.0
python-dotenv>=1.0.1,<2.0.0
python-json-logger>=2.0.7,<3.0.0
loguru>=0.7.0,<1.0.0
kubernetes>=30.1.0,<31.0.0
langchain>=0.3.0,<1.0.0
langchain-core>=0.3.0,<1.0.0
langgraph>=0.2.0,<1.0.0
groq>=0.11.0,<1.0.0
